Résolution dynamique et fix
Implémenter:
 - Implémenter Résolutions Dynamique √
 - Implémenter gestion de l'os différent selon si mac/linux avec dans option.os == darwin si mac √

Fix:
 - quitter State √
 - Retirer Decimal lors de l'affichage de la joie √
 - Augmenter la police des messages temp √
 - Perte de focus sur la carte et donc la perte du déplacement par touche fléché √
 - Différence Mac/Linux √
 - Transformer aipop.py en aipop.md √
 - SnowBale de la vassalisation de l'ennemie, Juste mettre une condition pour que l'ia ne puisse tenter qu'1 vassalisation par tour √
- Nettoyage Commentaire/Fonctions inutiles √

+++ functions/gameclass.py
@@ -334,7 +334,7 @@ class Classlord:
 			temp_joy += village.global_joy
 
 		if (len(self.vassal)+len(self.fief)) != 0:
-			self.global_joy = temp_joy/(len(self.vassal)+len(self.fief))
+			self.global_joy = int(temp_joy/(len(self.vassal)+len(self.fief)))
 		else:
 			self.global_joy = 0
 
@@ -659,7 +659,7 @@ class Classvillage:
 			self.prod_ressource += pop.tax_ressource()
 
 		# Calcul de la joie global du village
-		self.global_joy = temp_joy/len(self.population)
+		self.global_joy = int(temp_joy/len(self.population))
 
 
 	def endofturn(self, gamedata):
